requirements.txt
google-genai>=0.8.0
pinecone>=5.0.0
openai>=1.50.0
anthropic>=0.50.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
tqdm>=4.66.0
rich>=13.7.0
gradio>=4.40.0
pydantic>=2.5.0
pydantic-settings>=2.2.0
pypdf>=4.0.0
pymupdf>=1.23.0
